+++ saqc/__init__.py
@@ -9,12 +9,7 @@
 
 from saqc.version import __version__
 
-from saqc.constants import (
-    UNFLAGGED,
-    GOOD,
-    DOUBTFUL,
-    BAD,
-)
+from saqc.constants import UNFLAGGED, GOOD, DOUBTFUL, BAD, FILTER_NONE, FILTER_ALL
 
 # import order: from small to big, to a void cycles
 from saqc.core import (